fix mac and linux CI build

+++ SerialPrograms/Source/Pokemon/Inference/Pokemon_TrainIVCheckerOCR.h
@@ -32,7 +32,7 @@ public:
 private:
     StringOption DIRECTORY;
     EnumDropdownOption MODE;
-    SimpleIntegerOption<size_t> THREADS;
+    SimpleIntegerOption<unsigned int> THREADS;
 
 };
